Use better atan2 approximation; use float instead of double for faster computations; change LPF to 256 Hz and sample rate to 4 kHz; implement Complementary filter

+++ IMU/src/main.cpp
@@ -96,7 +96,7 @@ int main(void)
 	mpu->ready = true;
 
 	uint32_t ticks = Timer::Get_Tick_Count();
-	double roll, pitch, yaw;
+	float roll, pitch, yaw;
 
 	while (true) {
 		if (mpu->Data_Ready() && Timer::Get_Tick_Count() - ticks >= 1000000) {
